+++ wazo_phoned/plugins/fanvil/http.py
@@ -1,6 +1,7 @@
-# Copyright 2019-2020 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2019-2024 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
+from wazo_phoned.auth import AuthResource
 from wazo_phoned.plugin_helpers.client.http import ClientInput, ClientLookup
 
 
@@ -13,3 +14,25 @@ class Lookup(ClientLookup):
     MAX_ITEM_PER_PAGE = 16
     content_type = 'text/xml; charset=utf-8'
     template = 'fanvil_results.jinja'
+
+
+class DNDUserServiceEnable(AuthResource):
+    def __init__(self, service, *args, **kwargs):
+        super().__init__()
+        self._service = service
+
+    def get(self, user_uuid):
+        self._service.update_dnd(user_uuid, True)
+
+        return '', 200
+
+
+class DNDUserServiceDisable(AuthResource):
+    def __init__(self, service, *args, **kwargs):
+        super().__init__()
+        self._service = service
+
+    def get(self, user_uuid):
+        self._service.update_dnd(user_uuid, False)
+
+        return '', 200